Removes redundant check
"!c" was checked, and does not need to be checked again on 5866. 

After rewriting, the expression " !(c = nexttiled(c->next))" looks off to me. Is the assignment intended? If so, was the original value of c supposed to be called with "pop". It just seems odd that all the logic was base on checking "selmon->sel" (c's inital value), for us to then to perform an action on the result of value equal to "nexttiled(selmon->sel->next)".