Warp mouse to scratchpad only if focusfollowsmouse
When the scratchpad appears, the mouse is warped to the center of one of
the scratchpad windows. If the focus follows the mouse, this makes
sense, because the user is bringing up the scratchpad in order to
interact with it. However, if the focusfollowsmouse option is turned
off, the scratchpad can be focused without involving the mouse, so the
warp is disabled.