Rewrote `build.sh` and changed default behaviour
Foreword:
--------

After trying to change one or two things in `config.h`, I was displeased
when I saw that the old build script only performed four functions:

  1. Cleaned the working directory.
  2. Deleted `config.h` (thus prohibiting straight-forward, safe user
     modification)
  3. Compiled the WM source as a normal user
  4. Ran the install functions in the `Makefile`.

I sought to merely change the default behaviour to choosing whether or
not the preserve `config.h`.

New Default Behaviour:
---------------------

By default, if `config.h` exists, then it will not be deleted.  This is
the default behaviour of the script ran without any arguments or with
the `-c` or `--custom-config` arguments.  The checking of flags is done
via a case-statement.

This is not a user-facing feature, the first thing the script does is
checks whether the user is using the old `sudo` command, or the more
secure, newer `doas` command.  This is done by merely checking whether
both `/usr/bin/doas` and `/etc/doas.conf` exist.  If they exist, then
`doas` is used.  If they do not, then `sudo` is used.

Expressed simply, this would look like:

  1. Does the user mostlikely use doas?
  2. Clean working directory
  3. Recompile source as normal user
  4. Install

Old Default Behaviour:
---------------------

The old default behaviour is preserved.  It is accessible by passing the
`-D` or `--delete-config` flags.

Extended functionality:
----------------------

The scripts main body, which is not user-facing, are four functions
whose names are kennings.

  + default_behavior() - merely cleans working directory, runs `make` as
                         default user, then runs `make install` as root.
  + default_config()   - iteratively backs up `config.h`, then
                         recompiles using `config.def.h`.
  + delete_config()    - deletes `config.h` without backing up.
  + get_help()         - prints a user-friendly helper-message.

default_config():
----------------

Is the most important addition in this rewrite.  It looks like this:

    default_config() {
        if [[ -e config.h ]] && [[ ! -e config.h.orig ]]; then
            mv config.h config.h.orig
            make
            "$SUPERU" make install
        elif [[ -e config.h.orig ]] ; then
            i=0
            for i2 in config.h.orig* ; do
                let i++
            done
            mv config.h config.h.orig."$i"
            make
            "$SUPERU" make install
        else
            make
            "$SUPERU" make install
        fi
    }

It works by appending `.orig` to the first iteration `config.h` that is
to be backed up.  If `config.h.orig` exists, the second iteration will
have appended, `.1`.  If `config.h.orig.1` existis, then `2`, and so on.

The user is responsible for ensuring they do not have 5,000+ iterations.

Closing:
-------

`build.sh` now defaults to preserving `config.h`.  For testing purposes,
it is important that the old default behaviour of deleting it is
preserved.  The script may also intelligently create iterated backups.

All features have been extensively tested and pass, manually.

Automatically testing with `shellcheck build.sh` returned a warning
against an apparently "unused" counter, "$i2", which is vital for
creating iterations, so no error there as I understand (line 45).

It also recommended against using the "Bashism" "let i++" (line 46).

Please review the changes carefully before considering committing.